Extract declared propTypes detection code
