+++ lib/util/Components.js
@@ -10,11 +10,11 @@ const doctrine = require('doctrine');
 const variableUtil = require('./variable');
 const pragmaUtil = require('./pragma');
 const astUtil = require('./ast');
+const propTypes = require('./propTypes');
 
 function getId(node) {
   return node && node.range.join(':');
 }
-
 
 function usedPropTypesAreEquivalent(propA, propB) {
   if (propA.name === propB.name) {
@@ -695,12 +695,20 @@ function componentRule(rule, context) {
   // Update the provided rule instructions to add the component detection
   const ruleInstructions = rule(context, components, utils);
   const updatedRuleInstructions = util._extend({}, ruleInstructions);
-  Object.keys(detectionInstructions).forEach(instruction => {
+  const propTypesInstructions = propTypes.instructions(context, components, utils);
+  const allKeys = new Set(Object.keys(detectionInstructions).concat(Object.keys(propTypesInstructions)));
+  allKeys.forEach(instruction => {
     updatedRuleInstructions[instruction] = function(node) {
-      detectionInstructions[instruction](node);
+      if (instruction in detectionInstructions) {
+        detectionInstructions[instruction](node);
+      }
+      if (instruction in propTypesInstructions) {
+        propTypesInstructions[instruction](node);
+      }
       return ruleInstructions[instruction] ? ruleInstructions[instruction](node) : void 0;
     };
   });
+
   // Return the updated rule instructions
   return updatedRuleInstructions;
 }
